Prevent ssl disconnects
The self described hack doesn't seem to work in python 2.7, and I'm not
sure it's intended purpose. Whenever it gets hit, I get the error
'TypeError: argument of type 'int' is not iterable' and skybot stops
working.

I've removed it here and have had continuous uptime against an ssl irc
connection for a couple of weeks.