Join channels when config.json changes. Fixes #1
This uses the multi-channel join command, and appears to be harmless if
a client is already in a channel.

It does not part channels that aren't listed in its config, since the
bot responds to INVITEs by default. That could be implemented by
tracking the channel list (based on JOIN/PART messages received).

It may not work with very large channel lists.