Use yahoo finance CSV to retrieve stock info
This is the same underlying API that YQL uses.  While
YQL is a great abstractiong, it seems that many of the
YQL nodes do not work as expected - and either return
no results, very old results, or fail with various errors.

Using the CSV directly saves us from having to
blacklist "bad" YQL nodes.