ensure http util get decodes binary to utf-8
this definitely isn't "correct" but it's at least consistent
with the previous behavior.  To be correct, we should read
the suggested encoding from the HTTP content-type header
and attempt to decode the binary data to the suggested
charset.

also, update the other helpers to directly call `open` and `read`
without calling decode, because (in some cases, like lxml) they
cannot handle the unicode strings as expected in Python2