switch to another predb because orlydb is down
most of the predb sites seem to be using a very strict
cloudflare setting, so we can't scrap them.  PR3.us doesn't
but is definitely a bit on the weirder side to try to scrape.

if the accept headers are missing it seems to crash with the
underlying PHP, but at least it gives us the same data back as
with the other predb