Google Translate plugin without API Key access
this updates the google translate plugin to work the same as
the web application through a little bit of reverse engineering
the javascript.

the way this works is that we first retrieve an initialization
vector from translate.google.com, and then create a signature
token by signing the text inputs.

the original javascript for all of this was obfuscated before
rewriting it some in python.