simplify threads and drop @hook.singlethreaded
this makes it so all commands have their own thread, instead of spawning
endless threads that spin up and die.  This means a simplification of logic
when dealing with creating plugins and how they interact with the rest of
the bot.  this should benefit us because the bot isn't doing any sort of
management of a thread pool.