cleanup IRC connection hitting its own conf directly
the IRC connections were arbitrarily getting values from the conf,
for the connection and there was no source of truth for the default
values.  To clear things up, we're instead handling all of that within
the IRC classes themselves, and only reference the explicitly set values.

Also move some of the global regex to the IRC connection object to
clean up the global namespace